+++ src/dashboard/components/BrowserCompatibility.tsx
@@ -2,11 +2,11 @@
 import { ErrorOutline } from '@mui/icons-material';
 import { isWebGLSupported } from '@pixi/utils';
 import * as Sentry from '@sentry/browser';
+import { Empty } from 'components/Empty';
 import { Outlet } from 'react-router-dom';
-import { isWASMSupported } from '../../utils/isWASMSupported';
-import Empty from '../Empty';
+import { isWASMSupported } from 'utils/isWASMSupported';
 
-export default function BrowserCompatibility() {
+export function BrowserCompatibility() {
   if (!isWASMSupported || !isWebGLSupported()) {
     Sentry.captureEvent({
       message: 'Browser does not support WebGL or WASM',
