+++ src/dashboard/components/Header.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography, useTheme } from '@mui/material';
 import { ReactNode, useEffect } from 'react';
 
-export default function Header({ title, actions }: { title: string; actions?: ReactNode }) {
+export function Header({ title, actions }: { title: string; actions?: ReactNode }) {
   const theme = useTheme();
 
   useEffect(() => {
