+++ src/dashboard/files/ExamplesRoute.tsx
@@ -1,8 +1,8 @@
 import { EXAMPLE_FILES } from 'constants/app';
 import { ROUTES } from 'constants/routes';
+import { FileListItem } from 'dashboard/components/FileListItem';
+import { Header } from 'dashboard/components/Header';
 import { Link } from 'react-router-dom';
-import File from 'shared/dashboard/FileListItem';
-import Header from 'shared/dashboard/Header';
 
 export const Component = () => {
   return (
@@ -10,7 +10,7 @@ export const Component = () => {
       <Header title="Examples" />
       {Object.entries(EXAMPLE_FILES).map(([id, { name, description }]) => (
         <Link to={`${ROUTES.CREATE_FILE}?example=${id}`} style={{ textDecoration: 'none' }} key={id}>
-          <File name={name} description={description} />
+          <FileListItem name={name} description={description} />
         </Link>
       ))}
     </>
