+++ src/index.js
@@ -2,9 +2,9 @@ import * as Sentry from '@sentry/react';
 import { BrowserTracing } from '@sentry/tracing';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+import { Routes } from './Routes';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import { Routes } from './routes/index';
 import './styles.css';
 
 // Enable sentry only if SENTRY_DSN is in ENV
