+++ src/ui/contexts/File.tsx
@@ -1,10 +1,10 @@
 import { PostFileContentsReq, PostFileNameReq } from 'api-client/types';
+import { InitialFile } from 'dashboard/FileRoute';
 import { useDebounce } from 'hooks/useDebounce';
 import { useInterval } from 'hooks/useInterval';
 import mixpanel from 'mixpanel-browser';
 import { createContext, useCallback, useContext, useEffect, useRef, useState } from 'react';
 import { useParams } from 'react-router';
-import { InitialFile } from 'routes/file';
 import { GridFile } from 'schemas';
 import apiClientSingleton from '../../api-client/apiClientSingleton';
 import { SheetController } from '../../grid/controller/sheetController';
