+++ src/ui/menus/CellTypeMenu/CellTypeMenu.tsx
@@ -15,10 +15,10 @@ import React, { useCallback, useEffect } from 'react';
 import { useRouteLoaderData } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import inputRef from 'utils/inputRef';
+import { RootLoaderData } from '../../../Routes';
 import { editorInteractionStateAtom } from '../../../atoms/editorInteractionStateAtom';
 import { DOCUMENTATION_FORMULAS_URL, DOCUMENTATION_PYTHON_URL } from '../../../constants/urls';
 import { focusGrid } from '../../../helpers/focusGrid';
-import { RootLoaderData } from '../../../routes';
 import { CellType } from '../../../schemas';
 import { colors } from '../../../theme/colors';
 import { LinkNewTab } from '../../components/LinkNewTab';
