+++ src/ui/menus/CodeEditor/Console.tsx
@@ -3,10 +3,10 @@ import { useTheme } from '@mui/system';
 import { stripIndent } from 'common-tags';
 import { useEffect, useState } from 'react';
 import { useRouteLoaderData } from 'react-router-dom';
+import { RootLoaderData } from '../../../Routes';
 import { EditorInteractionState } from '../../../atoms/editorInteractionStateAtom';
 import { DOCUMENTATION_FORMULAS_URL, DOCUMENTATION_PYTHON_URL } from '../../../constants/urls';
 import { CellEvaluationResult } from '../../../grid/computations/types';
-import { RootLoaderData } from '../../../routes';
 import { Cell } from '../../../schemas';
 import { colors } from '../../../theme/colors';
 import { CodeSnippet } from '../../components/CodeSnippet';
