+++ src/ui/menus/ContextMenu/FloatingContextMenu.tsx
@@ -17,12 +17,12 @@ import { ControlledMenu, Menu, MenuItem, useMenuState } from '@szhsin/react-menu
 import mixpanel from 'mixpanel-browser';
 import { useCallback, useEffect, useRef } from 'react';
 import { GridInteractionState } from '../../../atoms/gridInteractionStateAtom';
+import { useGlobalSnackbar } from '../../../components/GlobalSnackbar';
 import { PNG_MESSAGE } from '../../../constants/app';
 import { copySelectionToPNG } from '../../../grid/actions/clipboard/clipboard';
 import { SheetController } from '../../../grid/controller/sheetController';
 import { PixiApp } from '../../../gridGL/pixiApp/PixiApp';
 import { KeyboardSymbols } from '../../../helpers/keyboardSymbols';
-import { useGlobalSnackbar } from '../../../shared/GlobalSnackbar';
 import { colors } from '../../../theme/colors';
 import { TooltipHint } from '../../components/TooltipHint';
 import { QColorPicker } from '../../components/qColorPicker';
