+++ src/ui/menus/TopBar/SubMenus/DataMenu.tsx
@@ -10,8 +10,8 @@ import { Tooltip } from '@mui/material';
 import Button from '@mui/material/Button';
 import { Menu, MenuHeader, MenuItem } from '@szhsin/react-menu';
 import '@szhsin/react-menu/dist/index.css';
+import { useGlobalSnackbar } from '../../../../components/GlobalSnackbar';
 import { CSV_IMPORT_MESSAGE } from '../../../../constants/app';
-import { useGlobalSnackbar } from '../../../../shared/GlobalSnackbar';
 import { MenuLineItem } from '../MenuLineItem';
 
 export const DataMenu = () => {
