+++ src/ui/menus/TopBar/SubMenus/QuadraticMenu.tsx
@@ -9,6 +9,7 @@ import { ROUTES } from 'constants/routes';
 import { useEffect } from 'react';
 import { useParams, useRouteLoaderData } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
+import { RootLoaderData } from '../../../../Routes';
 import { editorInteractionStateAtom } from '../../../../atoms/editorInteractionStateAtom';
 import { gridInteractionStateAtom } from '../../../../atoms/gridInteractionStateAtom';
 import { authClient } from '../../../../auth';
@@ -18,7 +19,6 @@ import { copyToClipboard, cutToClipboard, pasteFromClipboard } from '../../../..
 import { SheetController } from '../../../../grid/controller/sheetController';
 import { focusGrid } from '../../../../helpers/focusGrid';
 import { KeyboardSymbols } from '../../../../helpers/keyboardSymbols';
-import { RootLoaderData } from '../../../../routes';
 import { isMac } from '../../../../utils/isMac';
 import { MenuLineItem } from '../MenuLineItem';
 import { useGridSettings } from './useGridSettings';
